Separated generator script into separate components, added test to check generator exits w/o error
