+++ lerna-debug.log
@@ -1,0 +1,13 @@
+0 silly argv {
+0 silly argv   _: [ 'clean' ],
+0 silly argv   yes: true,
+0 silly argv   y: true,
+0 silly argv   lernaVersion: '3.22.1',
+0 silly argv   '$0': 'node_modules\\lerna\\cli.js'
+0 silly argv }
+1 notice cli v3.22.1
+2 verbose rootPath C:\Programming\user-office-lib
+3 info versioning independent
+4 error JSONError: Unexpected token } in JSON at position 1115 while parsing near '...ype": {	"module"  }}' in packages\uows-client-generator\package.json
+4 error     at module.exports (C:\Programming\user-office-lib\node_modules\@lerna\project\node_modules\parse-json\index.js:26:19)
+4 error     at parse (C:\Programming\user-office-lib\node_modules\@lerna\project\node_modules\load-json-file\index.js:15:9)
