+++ packages/uows-client-generator/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@stfc-user-programme/uows_client_generator",
-  "version": "1.0.9",
+  "version": "1.0.10",
   "description": "Generates an interface for consumning the User Office Web Service",
   "main": "lib/index.js",
   "author": "STFC",
